+++ source/plugin/carla-vst.cpp
@@ -749,7 +749,8 @@ protected:
 
     intptr_t handleDispatcher(const NativeHostDispatcherOpcode opcode, const int32_t index, const intptr_t value, void* const ptr, const float opt)
     {
-        carla_debug("NativePlugin::handleDispatcher(%i, %i, " P_INTPTR ", %p, %f)", opcode, index, value, ptr, opt);
+        carla_debug("NativePlugin::handleDispatcher(%i, %i, " P_INTPTR ", %p, %f)",
+                    opcode, index, value, ptr, static_cast<double>(opt));
 
         switch (opcode)
         {
